+++ src/main/java/com/hunteryavitz/blockchainapi/constants/ContaminationLevel.java
@@ -1,0 +1,32 @@
+package com.hunteryavitz.blockchainapi.constants;
+
+/**
+ * The ContaminationLevel enum represents the level of contamination of an exception.
+ */
+public enum ContaminationLevel {
+
+    /**
+     * The info level.
+     */
+    INFO,
+
+    /**
+     * The warning level.
+     */
+    WARNING,
+
+    /**
+     * The error level.
+     */
+    ERROR,
+
+    /**
+     * The critical level.
+     */
+    CRITICAL,
+
+    /**
+     * The unknown level.
+     */
+    UNKNOWN
+}