+++ src/main/java/com/hunteryavitz/blockchainapi/services/BlockchainService.java
@@ -15,6 +15,18 @@ import java.util.Arrays;
 public class BlockchainService {
 
     /**
+     * The health metric service.
+     */
+    HealthMetricService healthMetricService;
+
+    /**
+     * The constructor for the BlockchainService class.
+     */
+    public BlockchainService() {
+        healthMetricService = new HealthMetricService();
+    }
+
+    /**
      * The blockchain is an array of blocks.
      */
     private static Block[] blockchain;
@@ -25,14 +37,28 @@ public class BlockchainService {
     private static Boolean[] liveness;
 
     /**
+     * The blockCount is the number of blocks for the purposes of measuring production.
+     */
+    private static Integer[] contaminationSpectrum;
+
+    /**
      * The constructor for the BlockchainService class.
      */
     public void createInitialBlockchain() {
+
+        if (healthMetricService == null) {
+            healthMetricService = new HealthMetricService();
+        }
 
         liveness = new Boolean[100];
 
         blockchain = new Block[100];
         Block genesisBlock;
+
+        contaminationSpectrum = new Integer[5];
+        Arrays.fill(contaminationSpectrum, 0);
+
+        healthMetricService.resetBlockCount();
 
         try {
             long timestamp = System.currentTimeMillis();
@@ -43,6 +69,7 @@ public class BlockchainService {
         }
 
         blockchain[0] = genesisBlock;
+        healthMetricService.incrementBlockCount();
 
     }
 
@@ -70,6 +97,8 @@ public class BlockchainService {
                     System.currentTimeMillis(), "Block " + nextBlockIndex,
                     Utils.calculateHash(nextBlockIndex, previousBlock.getHash(),
                             System.currentTimeMillis(), "Block " + nextBlockIndex));
+
+            incrementBlockCount();
         } catch (NoSuchAlgorithmException e) {
             throw new RuntimeException(e);
         }
@@ -101,6 +130,7 @@ public class BlockchainService {
         }
 
         blockchain[nextBlockIndex] = block;
+        healthMetricService.incrementBlockCount();
     }
 
     /**
@@ -141,5 +171,31 @@ public class BlockchainService {
      */
     public void checkReadiness() {
         liveness = Utils.updateLiveness(liveness);
+    }
+
+    /**
+     * The getHealthMetrics method is responsible for returning the health metrics of the blockchain.
+     * @return The health metrics of the blockchain.
+     */
+    public Integer[] getHealthMetrics() {
+
+        // mock health metrics
+        contaminationSpectrum[0] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[1] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[2] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[3] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[4] = (int) (Math.random() * 100 + 1);
+
+        Integer[] healthMetrics = Utils.getContaminationSpectrum(contaminationSpectrum);
+        Arrays.fill(contaminationSpectrum, 0);
+
+         return healthMetrics;
+    }
+
+    /**
+     * The incrementBlockCount method is responsible for incrementing the block count.
+     */
+    private void incrementBlockCount() {
+        contaminationSpectrum[0]++;
     }
 }